Default args, whitespace, docstrings
The only functional change here is to call `pairwise_consts` only if
*both* `pσ` and `pϕ` are `nothing`. The implementation discards any
user-supplied values, and this could lead to surprising results if the
user supplies one but not the other. Accepting the error when you try to
index `nothing` seems like the safer approach, as it will force the user
to supply both or neither.

Aside from that, this just cleans up whitespace and fixes some
docstrings.