Provide more interface convenience functions (#47)
This fleshes out the interface of GMMs as vector- or set-like
(implementing `push!` and `pop!`), and MultiGMMs as dictionary-like
(implementing `valtype`, `get`, `get!`, `delete!`, and `empty!`). This
makes it easier to write code without needing to reach into internals.
The new interface functions are used to simplify some of the other code.