Eliminate piracy of CoordinateTransformations
This package extends several functions from CoordinateTransformations,
but it extends them in ways that do not involve types defined in this
package. Hence these methods are piracy.

`trans(A::AbstractMatrix)` could conceivably be defined in
CoordinateTransformations, so rather than merge this as-is it might be
better to consider submitting it there.