[RFC] Fix piracy (#51)
* Update to CoordinateTransformations.kabsch

`kabsch` has now been implemented in CoordinateTransformations. While
new features are ordinarily not breaking changes, for this package it is
due to name conflict. This PR resolves the name conflict, and bumps
`[compat]`.

Closes #49

* Eliminate piracy of CoordinateTransformations

This package extends several functions from CoordinateTransformations,
but it extends them in ways that do not involve types defined in this
package. Hence these methods are piracy.

`trans(A::AbstractMatrix)` could conceivably be defined in
CoordinateTransformations, so rather than merge this as-is it might be
better to consider submitting it there.

* Add Aqua tests

I've set these up to run only on CI, but if you prefer it's easier to
always run them. They just take a while.

* Eliminate piracy of CoordinateTransformations

This package extends several functions from CoordinateTransformations,
but it extends them in ways that do not involve types defined in this
package. Hence these methods are piracy.

`trans(A::AbstractMatrix)` could conceivably be defined in
CoordinateTransformations, so rather than merge this as-is it might be
better to consider submitting it there.

* Add Aqua tests

I've set these up to run only on CI, but if you prefer it's easier to
always run them. They just take a while.

* Pass Aqua tests

* Fix premature version bump

* Fix LinearAlgebra compat entry

* Switch to build_tform for local ICP methods

---------

Co-authored-by: Tom McGrath <tmcgrath325@gmail.com>